% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_Opt.R
\name{psi_Opt}
\alias{psi_Opt}
\alias{Psi_Opt}
\alias{rho_Opt}
\alias{psip_Opt}
\alias{wgt_Opt}
\title{Optimal Psi}
\usage{
psi_Opt(x, cc)

Psi_Opt(x, cc)

rho_Opt(x, cc)

psip_Opt(x, cc)

wgt_Opt(x, cc)
}
\arguments{
\item{x}{a numeric vector of points where the function will be evaluated.}

\item{cc}{a length 4 numeric vector: \code{cc[1]} is the tuning constant a,
\code{cc[2]} is the lower end point of the support interval \code{lower},
\code{cc[3]} is the upper end point of the support interval \code{upper},
and \code{cc[4]}}
}
\description{
Evaluate Yohai and Zamar's (1997) Optimal Psi function, its
             derivative, antiderivative, rho, and weight functions.
}
\details{
The family of optimal robust \eqn{\psi}{psi} functions is given by

         \deqn{\psi_{a}(x) = \mathrm{sign}(x) \max\left(|x| - \frac{a}{\phi(x)}, 0\right)}{%
                 psi(x, a) = sign(x)*max(abs(x) - a / dnorm(x), 0)}

         where \eqn{\phi(x)}{dnorm(x)} is the standard normal density function and
         \eqn{a} is the tuning parameter. For \eqn{x > 0}, \eqn{\psi_{a}(x)}{psi(x, a)}
         has support on the interval \eqn{(l_{a}, u_{a})}{(l, u)} where \eqn{l} and
         \eqn{u} depend on the value of \eqn{a} (see
         \code{\link{OptPsiSupportIntervalFromConst}}). Since the support interval
         is fixed for a given \eqn{a}, values for \eqn{l} and \eqn{u} are stored in
         the parameter vector \code{cc} to avoid unnecessary recomputation. Lastly,
         we introduce a scaling parameter \eqn{c}

         \deqn{\mathrm{psi_Opt(x, cc)} = \begin{cases} \mathrm{sign}(x) |x/c| - \frac{a}{\phi(x/c)} & |x/c| \in (l_{a}, u_{a}) \\ 0 & \mathrm{Otherwise}}{%
               psi(x, a, c) = sign(x/c)*max(abs(x/c) - a / dnorm(x/c), 0)}
}
\section{Functions}{
\itemize{
\item \code{Psi_Opt}: 

\item \code{rho_Opt}: 

\item \code{psip_Opt}: 

\item \code{wgt_Opt}: 
}}

